requirements.txt
aiml==0.9.2
aioredis==1.3.0
asgiref==3.2.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.7.0
blirupchat==1.0.0
certifi==2019.11.28
cffi==1.13.0
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
cli-helpers==1.2.1
Click==7.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
darr==0.2.2
dash==1.4.1
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.41.0
django-chatovod==1.4.2
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-countries==6.1.2
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-plotly-dash==1.0.1
django-pwa==1.0.10
django-redis==4.10.0
django-summernote==0.8.11.6
django-webcam==0.4.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dpd-components==0.1.0

Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.1
gunicorn==20.0.4
hiredis==1.0.0
humanize==2.0.0
hyperlink==19.0.0
idna==2.8
iexfinance==0.4.3
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.15.1
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
pandas==1.0.0
pgcli==2.2.0
pgspecial==1.11.9
Pillow==7.1.2
plotly==4.2.1
pocket==0.3.6
prompt-toolkit==2.0.10
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyFirmata==1.1.0
Pygments==2.6.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
PyQt5==5.14.2
PyQt5-sip==12.8.0
pyqt5-tools==5.14.2.1.7b3
pyserial==3.4
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2019.3
Qt.py==1.2.3
redis==3.3.11
reportlab==3.5.42
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
robodk==4.2.4
robotframework==3.1.2
scikit-learn==0.23.1
scipy==1.4.1
service-identity==18.1.0
setproctitle==1.1.10
six==1.12.0
sklearn==0.0
sound==0.1.0
SoundFile==0.10.3.post1
sqlparse==0.3.0
stripe==2.50.0
tabulate==0.8.6
terminaltables==3.1.0
threadpoolctl==2.0.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.8
uuid==1.30
wcwidth==0.1.8
Werkzeug==1.0.0
zope.interface==4.6.0
